requirements.txt
python-telegram-bot==13.15
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
numpy>=1.26.0
scikit-learn>=1.5.0
